+++ app/src/main/java/org/android/go/sopt/presentation/main/GalleryFragment.kt
@@ -1,17 +1,15 @@
-package org.android.go.sopt
+package org.android.go.sopt.presentation.main
 
-import MyAdapter
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import androidx.recyclerview.widget.LinearLayoutManager
-import org.android.go.sopt.databinding.FragmentHomeBinding
+import org.android.go.sopt.databinding.FragmentGalleryBinding
 
-class HomeFragment : Fragment() {
-    private var _binding: FragmentHomeBinding? = null
-    private val binding: FragmentHomeBinding
+class GalleryFragment : Fragment() {
+    private var _binding: FragmentGalleryBinding? = null
+    private val binding: FragmentGalleryBinding
         get() = requireNotNull(_binding) { "앗!_binding이 null이다!" }
 
     override fun onCreateView(
@@ -19,14 +17,12 @@ class HomeFragment : Fragment() {
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View {
-        _binding = FragmentHomeBinding.inflate(inflater, container, false)
+        _binding = FragmentGalleryBinding.inflate(inflater, container, false)
         return binding.root
     }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
-        binding.rvHomeGithubRepo.adapter = MyAdapter(requireContext())
-        binding.rvHomeGithubRepo.layoutManager = LinearLayoutManager(context)
     }
 
     override fun onDestroy() {
