+++ app/src/main/java/org/android/go/sopt/presentation/main/HomeFragment.kt
@@ -1,15 +1,16 @@
-package org.android.go.sopt
+package org.android.go.sopt.presentation.main
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import org.android.go.sopt.databinding.FragmentGalleryBinding
+import androidx.recyclerview.widget.LinearLayoutManager
+import org.android.go.sopt.databinding.FragmentHomeBinding
 
-class GalleryFragment : Fragment() {
-    private var _binding: FragmentGalleryBinding? = null
-    private val binding: FragmentGalleryBinding
+class HomeFragment : Fragment() {
+    private var _binding: FragmentHomeBinding? = null
+    private val binding: FragmentHomeBinding
         get() = requireNotNull(_binding) { "앗!_binding이 null이다!" }
 
     override fun onCreateView(
@@ -17,12 +18,14 @@ class GalleryFragment : Fragment() {
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View {
-        _binding = FragmentGalleryBinding.inflate(inflater, container, false)
+        _binding = FragmentHomeBinding.inflate(inflater, container, false)
         return binding.root
     }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
+        binding.rvHomeGithubRepo.adapter = MyAdapter(requireContext())
+        binding.rvHomeGithubRepo.layoutManager = LinearLayoutManager(context)
     }
 
     override fun onDestroy() {
