+++ app/src/main/java/org/android/go/sopt/presentation/main/SearchFragment.kt
@@ -1,4 +1,4 @@
-package org.android.go.sopt
+package org.android.go.sopt.presentation.main
 
 import android.content.Context
 import android.os.Bundle
