+++ app/src/main/res/layout/item_github_repo.xml
@@ -5,7 +5,7 @@
     android:layout_height="wrap_content"
     android:padding="10dp">
 
-    <androidx.constraintlayout.utils.widget.ImageFilterView
+    <ImageView
         android:layout_width="100dp"
         android:layout_height="0dp"
         android:src="@drawable/img_introduce"
